Implementado o player principal de áudio na playlist e realizados os primeiros testes de reprodução
