+++ public/index.html
@@ -2,14 +2,14 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="icon" href="favicon.png" />
+    <link rel="icon" href="/public/images/icons/favicon.png"/>
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
       content="Spotify Web site clone using create-react-app"
     />
-    <title>Escutar muda tudo - Spotify</title>
+    <title>Escutar muda tudo | Spotify</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
