+++ src/components/Container/index.js
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 export const Container = styled.div`
     display: grid;
     grid-template-columns: 1fr 4fr;
-    grid-template-rows: 100vh;
+    /* grid-template-rows: 100vh; */
     height: 100vh;
     
 `;
