+++ src/old/_App.css
@@ -1,0 +1,87 @@
+.App {
+  -webkit-box-pack: center;
+  display: flex;
+  flex-direction: column;
+  padding-left: 0px;
+  padding-right: 0px;
+  justify-content: space-between;
+  min-height: 846px;
+}
+
+.backgroundHome {
+  background-image: url(https://www.scdn.co/i/home/lieVariants/bursts.svg);
+  background-position-x: 46%;
+  background-position-y: 4%;
+  background-size: 175%;
+  background-repeat: no-repeat;
+  background-color: #2D46B9;
+}
+
+.navbar {
+  height: 80px;
+}
+
+.row {
+  margin-right: 0;
+  margin-left: 0;
+}
+
+.Menu {
+  background-color: black;
+  position: sticky;
+  display: block;
+  width: 100%;
+  left:0;
+  top:0;
+  z-index: 10;
+}
+
+.containerDefault {
+  height: 100vh;
+}
+
+.Rodape {
+  background-color: black;
+  display: block;
+  width: 100%;  
+  left:0;
+  bottom: 0;
+}
+
+.ConteudoHome {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-top: 10%;
+}
+
+.ConteudoHome h4 {
+  color: #1ED760;
+  font-size: 18px;
+}
+
+.backgroundFaq {
+  background-image: url(https://support.scdn.co/web/_next/static/assets/ef7fb47/search-desktop.webp);
+  background-repeat: no-repeat;
+  background-color: white;
+}
+
+.ConteudoFaq {
+  color: white;
+}
+
+.card-header{
+  background-color: black;
+}
+
+.card-header button {
+   color: white;
+}
+
+.card-body{
+  color: black;
+}
+
+.card{
+  border-style: none;
+}
