--- src/view/Help/styles.css
@@ -1,28 +1,0 @@
-.ConteudoFaq{
-    height: 100vh;
-    
-}
-
-.ConteudoFaq {
-    color: white;
-  }
-  
-  .card-header{
-    background-color: black;
-  }
-  
-  .card-header button {
-     color: white;
-  }
-  
-  .card-body{
-    color: black;
-  }
-  
-  .card{
-    border-style: none;
-}
-
-a{
-    text-decoration: none;
-}
